+++ src/auth/uriHandler.ts
@@ -6,6 +6,10 @@ export class UriEventHandler extends vscode.EventEmitter<vscode.Uri> implements 
 			vscode.commands.executeCommand('vscode-mosayic.signIn');
 			return;
 		}
+		if (uri.path === '/wake') {
+			vscode.commands.executeCommand('vscode-mosayic.connect');
+			return;
+		}
 		this.fire(uri);
 	}
 }